motor-controller/requirements.txt
RPi.GPIO==0.7.0
python-dotenv==0.19.0